requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.23
python-dotenv==0.15.0
pandas~=1.2.3
progressbar2
scikit-learn~=0.24.1